+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - The sample transcript will now display the target for scoring in the Score Event (for newly run evaluations).
 - Provide setter for `max_messages` on `TaskState`. 
 - Provide `max_messages` option for `basic_agent()` (defaulting to 50) and use it rather than any task `max_messages` defined.
+- Improve quality of error messages when a model API key environment variable is missing.
 
 ## v0.3.38 (3 October 2024)